fix(proxy): offload openai/gemini token counting
The OpenAI and Gemini handlers resolved the tokenizer and counted messages
inline on the event loop. For an HF-routed model (qwen, deepseek, llama) a
cold cache runs AutoTokenizer.from_pretrained behind a 10s thread.join and
freezes the whole server (GH #1701); OpenAI /chat and /responses are
multi-provider passthroughs, so they receive those models.

Hoist the fail-open _count_tokens_offloaded helper from AnthropicHandlerMixin
to the shared HeadroomProxy base and route all six sites through it. It runs
resolution + counting on the bounded compression executor and fails open to
estimation on timeout, error, or quarantine, so an unrelated compression
timeout can't 500 a request that is only counting tokens.
