+++ headroom/proxy/handlers/anthropic.py
@@ -71,48 +71,6 @@ def _strip_index_from_content_blocks(content: Any) -> None:
 class AnthropicHandlerMixin:
     """Mixin providing Anthropic API handler methods for HeadroomProxy."""
 
-    async def _count_tokens_offloaded(self, model, messages):  # noqa: ANN001, ANN201
-        """Resolve a tokenizer and count messages off the event loop.
-
-        Tokenizer resolution can be expensive on first use (HuggingFace
-        backends may download vocab files) and counting a full Claude Code
-        conversation is CPU-bound, so both run on the compression executor
-        bounded by ``COMPRESSION_TIMEOUT_SECONDS`` (GH #1701: an unbounded
-        on-loop load froze the whole server). On timeout or error this
-        fails open to character-based estimation.
-
-        Returns:
-            Tuple of ``(tokenizer, token_count)``. The tokenizer is fully
-            initialized, so later ``count_messages`` calls on it are pure
-            CPU work.
-        """
-        from headroom.proxy.helpers import COMPRESSION_TIMEOUT_SECONDS
-        from headroom.tokenizers import EstimatingTokenCounter, get_tokenizer
-
-        def _resolve_and_count():  # noqa: ANN202
-            tokenizer = get_tokenizer(model)
-            return tokenizer, tokenizer.count_messages(messages)
-
-        try:
-            return await self._run_compression_in_executor(
-                _resolve_and_count,
-                timeout=float(COMPRESSION_TIMEOUT_SECONDS),
-            )
-        except Exception as e:  # fail open — includes asyncio.TimeoutError
-            # Log the downgrade once per model, not per request.
-            fallback_models = getattr(self, "_token_count_fallback_models", None)
-            if fallback_models is None:
-                fallback_models = set()
-                self._token_count_fallback_models = fallback_models
-            if model not in fallback_models:
-                fallback_models.add(model)
-                logger.warning(
-                    f"Token counting for model {model} failed or timed out "
-                    f"({e.__class__.__name__}); falling back to estimation"
-                )
-            estimator = EstimatingTokenCounter()
-            return estimator, estimator.count_messages(messages)
-
     @staticmethod
     def _resolve_ccr_workspace(
         request: Any,
